requirements.txt
anyio==3.6.2
asyncpraw==7.6.0
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.87.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
loguru==0.6.0
lxml==4.9.1
nlpcloud==1.0.34
oauthlib==3.2.2
openai==0.25.0
progress==1.6
pydantic==1.10.2
python-dateutil==2.8.2
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
Reverso-API==0.0.1b3
searchtweets==1.7.6
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.21.0
tqdm==4.64.1
tweet-parser==1.13.2
twitter-ads==11.0.0
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
nltk==3.7
emoji==2.2.0
deepl==1.13.0
numpy==1.23.4
databases==0.6.2
aiosqlite==0.17.0
sshtunnel==0.4.0
psycopg2-binary==2.9.5
asyncpg==0.27.0
python-multipart==0.0.5
httpx==0.23.1
Pillow==9.3.0
replicate==0.4.0
celery==5.2.7
redislite==6.2.805324
websockets==10.4
Jinja2==3.1.2